Make docs compatible
