+++ .storybook/preview.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { Provider as EasyUIProvider } from "../easy-ui-react/src/Provider";
 import { viewports } from "./viewports";
+import { theme } from "./theme";
 
 import "./poppins.css";
 
@@ -13,6 +14,9 @@ export const parameters = {
   },
   viewport: {
     viewports,
+  },
+  docs: {
+    theme,
   },
 };
