+++ easy-ui-react/src/Icon/Icon.stories.mdx
@@ -7,7 +7,7 @@ import WarningIcon from "@easypost/easy-ui-icons/Warning";
 import ErrorIcon from "@easypost/easy-ui-icons/Error";
 
 <Meta
-  title="Icon"
+  title="Components/Icon"
   argTypes={{
     // TODO: Figure out how to include all icons from our icons project
     symbol: createLabelledOptionsControl({
