+++ easy-ui-react/src/Placeholder/Placeholder.stories.mdx
@@ -3,7 +3,7 @@ import { Placeholder } from "./Placeholder";
 import { createLabelledOptionsControl } from "../utilities/storybook";
 
 <Meta
-  title="Placeholder"
+  title="Components/Placeholder"
   argTypes={{
     message: { control: "text" },
     iconSize: createLabelledOptionsControl({
